requirements-service.txt
"""
Service dependencies for HunterTrace API.
Install with: pip install -r requirements-service.txt
"""
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
PyYAML>=6.0
